requirements.txt
streamlit
pandas
numpy
Pillow
opencv-python-headless==4.8.1.78
pytesseract
reportlab
streamlit-drawable-canvas
scikit-learn
